refactor: move tests to external directory, add library entry point, and implement security warning for public interface binding
